Prevent possible backtracking index overflow on 64-bit platforms
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@38608 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/algo/align/nw/nw_band_aligner.hpp
@@ -83,8 +83,8 @@ protected:
     // band width
     size_t   m_band;
 
-    // band shift along the first sequences (can be negative)
-    int      m_Shift;
+    // band shift along the first sequence (can be negative)
+    long     m_Shift;
 
     // backtrace helpers
     size_t   m_TermK;
@@ -100,7 +100,7 @@ protected:
 
     // other
     void x_CheckParameters(const SAlignInOut* data) const;
-    virtual bool   x_CheckMemoryLimit(void);
+    virtual bool x_CheckMemoryLimit(void);
 };
